chore: remove CLI entry guard
