+++ src/index.ts
@@ -1,17 +1,8 @@
-#!/usr/bin/env node
-import { Command } from 'commander';
-import { writeFileSync, statSync } from 'node:fs';
-import path from 'node:path';
-import { fileURLToPath } from 'node:url';
 import { DiagramService } from './core/usecases/generateDiagram.js';
 import { TypeScriptParser } from './infrastructure/parsers/typescriptParser.js';
 import { MermaidDiagramGenerator } from './infrastructure/diagram/mermaidGenerator.js';
 import { LspParser } from './infrastructure/parsers/lspParser.js';
 import { StdioLanguageClient } from './infrastructure/lsp/stdioClient.js';
-
-const program = new Command();
-
-program.name('docgram').description('Generate diagrams from source code');
 
 export function buildService(parserOption: string) {
   const generator = new MermaidDiagramGenerator();
@@ -25,36 +16,6 @@ export function buildService(parserOption: string) {
   return new DiagramService(parser, generator);
 }
 
-program
-  .command('diagram')
-  .argument('<path...>', 'File or directory to parse')
-  .option('--parser <parser>', 'Parser implementation to use (ts|lsp)', 'ts')
-  .action(async (paths: string[], opts) => {
-    const service = buildService(opts.parser);
-    const diagram = await service.generateFromPaths(paths);
-    console.log(diagram);
-  });
-
-program
-  .command('docs')
-  .argument('<path...>', 'File or directory to parse')
-  .option('--parser <parser>', 'Parser implementation to use (ts|lsp)', 'ts')
-  .action(async (paths: string[], opts) => {
-    const service = buildService(opts.parser);
-    const diagram = await service.generateFromPaths(paths);
-    const target = paths[0];
-    const dir = statSync(target).isDirectory() ? target : path.dirname(target);
-    const title = path.basename(dir);
-    const content = `# ${title}\n\n\`\`\`mermaid\n${diagram}\n\`\`\`\n`;
-    const readmePath = path.join(dir, 'README.md');
-    writeFileSync(readmePath, content);
-    console.log(`README written to ${readmePath}`);
-  });
-
-if (fileURLToPath(import.meta.url) === process.argv[1]) {
-  program.parse();
-}
-
 export {
   DiagramService,
   TypeScriptParser,
@@ -62,3 +23,4 @@ export {
   LspParser,
   StdioLanguageClient,
 };
+
